refactor: 拆分播放器内部状态职责
