--- native/audio-engine/src/player.rs
@@ -1,1059 +1,0 @@
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::Arc;
-use std::thread::{self, JoinHandle};
-use std::time::Duration;
-
-use anyhow::Result;
-use ffmpeg_audio::HttpCancelHandle;
-use parking_lot::Mutex;
-use rodio::Player as RodioPlayer;
-use tracing::{debug, info};
-
-use crate::audio_output::AudioOutput;
-use crate::decoder;
-use crate::equalizer::{Equalizer, EQ_BAND_COUNT};
-use crate::fft::FftAnalyzer;
-use crate::shared::{AudioMetadata, Shared};
-use crate::source::DecoderSource;
-use crate::tempo::StretchProcessor;
-
-/// 播放器推送给 JS 侧的事件类型
-#[derive(Clone, Debug)]
-pub enum PlayerEvent {
-    /// 状态变化
-    StateChanged { state: PlayerState },
-    /// 播放结束
-    Ended,
-    /// 音源失效（网络中断 / URL 过期）
-    SourceError,
-    /// 位置更新（秒）—— 由内部定时器推送
-    Position { position: f64, duration: f64 },
-    /// FFT 频谱数据推送
-    FftData { ldata: Vec<f32>, rdata: Vec<f32> },
-    /// 输出流停滞（rodio sink 长时间未消费样本，需要外部重建输出）
-    OutputStalled,
-}
-
-/// 事件发射器类型（跨线程安全）
-pub type EventEmitter = Arc<dyn Fn(PlayerEvent) + Send + Sync>;
-
-/// 根据解码结束原因与当前位置决定对外完成事件
-fn playback_completion_event(shared: &Shared, duration: f64, position: f64) -> PlayerEvent {
-    let mid_stream = duration <= 0.0 || duration - position > 3.0;
-    if shared.is_decode_failed() && mid_stream {
-        PlayerEvent::SourceError
-    } else {
-        PlayerEvent::Ended
-    }
-}
-
-/// 渐变步数
-const FADE_STEPS: u32 = 20;
-
-/// 可取消的渐变：在独立线程中逐步调整音量，cancel 为 true 时提前退出
-fn fade_volume(sink: &RodioPlayer, from: f32, to: f32, duration_ms: u64, cancel: &AtomicBool) {
-    if duration_ms == 0 {
-        sink.set_volume(to);
-        return;
-    }
-    let step_duration = Duration::from_millis(duration_ms / u64::from(FADE_STEPS));
-    for step in 1..=FADE_STEPS {
-        if cancel.load(Ordering::Relaxed) {
-            return;
-        }
-        let progress = step as f32 / FADE_STEPS as f32;
-        sink.set_volume(from + (to - from) * progress);
-        // 分片可取消：渐变时长用户可配，长渐变的整步 sleep 会让 cancel_fade 的
-        // 同步 join 卡住最长一个步长
-        sleep_unless_stopped(cancel, step_duration);
-    }
-}
-
-/// 播放状态
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum PlayerState {
-    Idle,
-    Playing,
-    Paused,
-    Stopped,
-}
-
-/// 内部播放器，管理音频输出、解码和状态
-pub struct InnerPlayer {
-    /// 跨线程安全的音频输出包装：内部专用线程独占 cpal::Stream，
-    /// 此处只持有 Send 的 Mixer，保证 InnerPlayer 整体是 Send 的
-    output: Option<AudioOutput>,
-    /// 使用 Arc 包装，允许 fade 线程在 Mutex 外操作音量
-    sink: Option<Arc<RodioPlayer>>,
-    shared: Option<Arc<Shared>>,
-    /// 解码线程句柄，join 后可回收 DecoderData 复用于 seek
-    decoder_thread: Option<JoinHandle<decoder::DecoderData>>,
-    fft: Arc<FftAnalyzer>,
-    /// 当前音频的采样率（seek 重建 DecoderSource 时需要）
-    audio_sample_rate: u32,
-    /// 当前音频的声道数
-    audio_channels: u16,
-    /// 当前音频的时长（秒）
-    audio_duration: f64,
-    /// 原始封面数据缓存（load 时提取，getCoverRaw 时返回，避免重复打开文件）
-    cover_raw: Option<Vec<u8>>,
-    state: PlayerState,
-    /// seek 偏移基准（秒），采样计数在此基础上累加
-    seek_base: f64,
-    /// 当前音频源路径/地址
-    current_source: Option<String>,
-    /// 用户设置的目标音量（fade 期间 sink 音量会变化，需要记住目标值）
-    target_volume: f32,
-    /// 渐变时长（毫秒），0 表示禁用
-    fade_duration_ms: u64,
-    /// 封面缓存目录
-    cover_cache_dir: Option<String>,
-    /// 事件回调（由 lib.rs 设置，内部转发到 JS ThreadsafeFunction）
-    event_callback: Option<EventEmitter>,
-    /// 位置推送定时器的停止信号和线程句柄
-    position_timer_stop: Option<Arc<AtomicBool>>,
-    position_timer_handle: Option<JoinHandle<()>>,
-    /// 渐变取消信号和线程句柄
-    fade_cancel: Option<Arc<AtomicBool>>,
-    fade_handle: Option<JoinHandle<()>>,
-    /// FFT 推送开关（前端需要显示频谱时才启用）
-    fft_enabled: Arc<AtomicBool>,
-    /// FFT 推送定时器的停止信号和线程句柄
-    fft_timer_stop: Option<Arc<AtomicBool>>,
-    fft_timer_handle: Option<JoinHandle<()>>,
-    /// 用户选择的输出设备名称（None = 系统默认）
-    selected_device_name: Option<String>,
-    /// 音量归一化开关
-    normalization_enabled: bool,
-    /// 跨曲目共享的均衡器（load/seek 时交给 DSP 线程）
-    equalizer: Arc<Mutex<Equalizer>>,
-    /// 跨曲目共享的变速变调处理器（load/seek 时交给 DSP 线程，
-    /// set_speed/set_pitch/set_pitch_sync 直接锁此字段更新参数）
-    tempo: Arc<Mutex<StretchProcessor>>,
-    /// load 单调递增 token：每次 take_for_async_load 自增一次
-    /// commit_loaded 比对 token 与最新值，不一致则该次加载已被新加载取代，需丢弃
-    /// 用于防止快速切歌时旧 IO 完成后覆盖新音频的竞态
-    load_token: Arc<AtomicU64>,
-    /// 正在打开的网络音源中断句柄，确保切歌和 stop 能取消元数据探测
-    pending_load_handle: Option<HttpCancelHandle>,
-}
-
-/// 切换/seek 时要 join 的旧线程集合，全部挪到 spawn_blocking 工作线程 join，
-/// 主线程持锁阶段只 take handle，避免最坏 200ms+ 的卡顿
-pub struct OldThreads {
-    pub decoder_thread: Option<JoinHandle<decoder::DecoderData>>,
-    pub position_timer: Option<JoinHandle<()>>,
-    pub fft_timer: Option<JoinHandle<()>>,
-    pub fade_handle: Option<JoinHandle<()>>,
-}
-
-impl OldThreads {
-    /// 在工作线程上 join 所有旧 timer/fade，返回旧解码线程 handle 供调用方继续使用
-    /// 忽略 join 错误：辅助线程 panic 不阻止新加载，主播放路径不依赖它们
-    pub fn join_aux(self) -> Option<JoinHandle<decoder::DecoderData>> {
-        for h in [self.position_timer, self.fft_timer, self.fade_handle]
-            .into_iter()
-            .flatten()
-        {
-            let _ = h.join();
-        }
-        self.decoder_thread
-    }
-}
-
-/// async seek 阶段 1 的输出：带到工作线程做 join + ffmpeg seek + 重启解码
-pub struct SeekTake {
-    /// 所有旧线程 handle（工作线程 join）
-    pub old_threads: OldThreads,
-    /// 归一化开关（继承到新 Shared）
-    pub normalization_enabled: bool,
-    /// 归一化增益（继承到新 Shared）
-    pub normalization_gain: f32,
-    /// 当前音频源（seek 失败时 fallback 到 load）
-    pub current_source: Option<String>,
-    /// seek 前是否在播放（fallback 到 load 时保留状态）
-    pub was_playing: bool,
-    /// 当前输出设备采样率（新 Shared 沿用，与复用的重采样器目标一致）
-    pub output_sample_rate: u32,
-    /// 本次 seek 的 token，commit_seeked 时比对最新值，不一致说明已被新 load/seek/stop 取代
-    pub token: u64,
-    /// 解码侧 DSP 共享实例
-    pub equalizer: Arc<Mutex<Equalizer>>,
-    pub tempo: Arc<Mutex<StretchProcessor>>,
-}
-
-/// 完成音源准备后一次性提交给播放器的资源
-pub struct LoadedPlayback {
-    pub metadata: AudioMetadata,
-    pub decode_handle: JoinHandle<decoder::DecoderData>,
-    pub shared: Arc<Shared>,
-    pub output: AudioOutput,
-    pub cancel: Option<HttpCancelHandle>,
-}
-
-/// 编译期保证 `InnerPlayer: Send`：cpal::Stream（!Send）已通过 AudioOutput 隔离到专用线程，
-/// 此处不再需要 `unsafe impl Send`。如果未来有人加了 !Send 字段，这条断言会编译失败提醒。
-const _: fn() = || {
-    fn assert_send<T: Send>() {}
-    assert_send::<InnerPlayer>();
-};
-
-/// 分片 sleep：每 10ms 检查一次停止标志
-/// 让 stop_*_timer 的同步 join（pause/stop 都在 NAPI 主线程调用）在 ~10ms 内返回，
-/// 而不是阻塞整个推送周期（200ms）
-fn sleep_unless_stopped(flag: &AtomicBool, total: Duration) {
-    const SLICE: Duration = Duration::from_millis(10);
-    let mut remaining = total;
-    while !flag.load(Ordering::Relaxed) && !remaining.is_zero() {
-        let step = remaining.min(SLICE);
-        thread::sleep(step);
-        remaining -= step;
-    }
-}
-
-impl InnerPlayer {
-    /// 未初始化时通过 `AudioOutput::new` 懒构造音频输出。
-    /// 设备失效时的重建由 `reinit_output` 显式处理，不在此函数内自动恢复
-    fn ensure_output(&mut self) -> Result<&AudioOutput> {
-        if self.output.is_none() {
-            self.output = Some(AudioOutput::new(self.selected_device_name.as_deref())?);
-        }
-        self.output
-            .as_ref()
-            .ok_or_else(|| anyhow::anyhow!("ensure_output 后置条件违反"))
-    }
-
-    /// 当前实际输出流采样率（播放重采样目标）
-    pub fn output_sample_rate(&self) -> u32 {
-        self.output
-            .as_ref()
-            .map(|out| out.sample_rate())
-            .unwrap_or(decoder::DEFAULT_TARGET_SAMPLE_RATE)
-    }
-
-    pub fn new() -> Result<Self> {
-        // 延迟初始化：构造时不要求有音频设备，load 时再打开
-        let output = None;
-        let initial_rate = decoder::DEFAULT_TARGET_SAMPLE_RATE;
-        debug!("InnerPlayer 已创建");
-
-        Ok(Self {
-            output,
-            sink: None,
-            shared: None,
-            decoder_thread: None,
-            fft: Arc::new(FftAnalyzer::new()),
-            audio_sample_rate: 0,
-            audio_channels: 0,
-            audio_duration: 0.0,
-            cover_raw: None,
-            state: PlayerState::Idle,
-            seek_base: 0.0,
-            current_source: None,
-            target_volume: 1.0,
-            fade_duration_ms: 200,
-            cover_cache_dir: None,
-            event_callback: None,
-            position_timer_stop: None,
-            position_timer_handle: None,
-            fade_cancel: None,
-            fade_handle: None,
-            fft_enabled: Arc::new(AtomicBool::new(false)),
-            fft_timer_stop: None,
-            fft_timer_handle: None,
-            selected_device_name: None,
-            normalization_enabled: false,
-            equalizer: Arc::new(Mutex::new(Equalizer::new(initial_rate))),
-            tempo: Arc::new(Mutex::new(StretchProcessor::new(
-                decoder::TARGET_CHANNELS,
-                initial_rate,
-            ))),
-            load_token: Arc::new(AtomicU64::new(0)),
-            pending_load_handle: None,
-        })
-    }
-
-    /// 切换输出设备（下一次重建设备时生效）
-    pub fn set_output_device(&mut self, device_name: Option<String>) {
-        info!(device = ?device_name, "切换输出设备");
-        self.selected_device_name = device_name;
-    }
-
-    /// 获取当前选择的输出设备名称（None = 系统默认）
-    pub fn selected_device_name(&self) -> Option<&str> {
-        self.selected_device_name.as_deref()
-    }
-
-    /// 获取当前音频源
-    pub fn current_source(&self) -> Option<String> {
-        self.current_source.clone()
-    }
-
-    /// 注册事件回调（支持热替换：先停止旧的定时器/渐变，确保旧回调的 Arc 引用尽快释放）
-    pub fn set_event_callback(&mut self, cb: EventEmitter) {
-        self.stop_position_timer();
-        self.stop_fft_timer();
-        self.cancel_fade();
-        self.event_callback = Some(cb);
-    }
-
-    /// 发射事件
-    fn emit(&self, event: PlayerEvent) {
-        if let Some(cb) = &self.event_callback {
-            cb(event);
-        }
-    }
-
-    /// 对外发 SourceError：供 lib.rs 在远端源内部重开（seek 回退 / play 复活）失败时通知 JS 重解析
-    pub fn emit_source_error(&self) {
-        self.emit(PlayerEvent::SourceError);
-    }
-
-    /// 取消正在进行的渐变，并等待渐变线程退出
-    fn cancel_fade(&mut self) {
-        if let Some(flag) = self.fade_cancel.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(handle) = self.fade_handle.take() {
-            let _ = handle.join();
-        }
-    }
-
-    /// 启动非阻塞渐变（独立线程执行，不阻塞调用方）。
-    /// 完成回调仅在未被取消时执行
-    fn start_fade(&mut self, from: f32, to: f32, on_complete: Option<Box<dyn FnOnce() + Send>>) {
-        self.cancel_fade();
-
-        let cancel = Arc::new(AtomicBool::new(false));
-        self.fade_cancel = Some(Arc::clone(&cancel));
-
-        if let Some(ref sink) = self.sink {
-            let sink = Arc::clone(sink);
-            let fade_ms = self.fade_duration_ms;
-            let handle = thread::spawn(move || {
-                fade_volume(&sink, from, to, fade_ms, &cancel);
-                if !cancel.load(Ordering::Relaxed) {
-                    if let Some(callback) = on_complete {
-                        callback();
-                    }
-                }
-            });
-            self.fade_handle = Some(handle);
-        }
-    }
-
-    /// 启动位置推送定时器（在独立线程中运行，每 200ms 推送一次位置）
-    fn start_position_timer(&mut self) {
-        self.stop_position_timer();
-
-        let stop_flag = Arc::new(AtomicBool::new(false));
-        self.position_timer_stop = Some(Arc::clone(&stop_flag));
-
-        let shared = match &self.shared {
-            Some(s) => Arc::clone(s),
-            None => return,
-        };
-        let cb = match &self.event_callback {
-            Some(cb) => Arc::clone(cb),
-            None => return,
-        };
-        let seek_base = self.seek_base;
-        let duration = self.duration();
-
-        let handle = thread::spawn(move || {
-            // 停滞检测：消费计数连续 STALL_THRESHOLD_TICKS 次未变 + 缓冲区非空 → sink 静默死亡
-            const STALL_THRESHOLD_TICKS: u32 = 6; // 6 * 200ms = 1.2s
-            let mut last_consumed = shared.samples_consumed_count();
-            let mut stall_ticks: u32 = 0;
-
-            while !stop_flag.load(Ordering::Relaxed) {
-                let consumed = shared.samples_consumed_count();
-                let position = seek_base + shared.consumed_position();
-                cb(PlayerEvent::Position { position, duration });
-
-                // 检测播放结束：all_consumed 表示 rodio 侧已消费完所有数据
-                if shared.is_all_consumed() {
-                    // 解码因读取失败中止且距末尾尚远 → 音源失效，前端重新解析地址续播；
-                    // 距末尾 3s 内的失败按正常结束处理——Content-Length 偏大的转码源
-                    // 在曲尾必然提前 EOF，整曲重载只会带来一轮无意义抖动。
-                    // duration 未知（直播流等）时无"末尾"概念，失败一律上报
-                    cb(playback_completion_event(&shared, duration, position));
-                    cb(PlayerEvent::StateChanged {
-                        state: PlayerState::Stopped,
-                    });
-                    break;
-                }
-
-                // 缓冲区空时认为是解码 underrun，等待解码补数据，不视为停滞
-                if consumed == last_consumed && !shared.is_buffer_empty() {
-                    stall_ticks += 1;
-                    if stall_ticks >= STALL_THRESHOLD_TICKS {
-                        cb(PlayerEvent::OutputStalled);
-                        // 发完归零，依赖主进程冷却防抖
-                        stall_ticks = 0;
-                    }
-                } else {
-                    stall_ticks = 0;
-                }
-                last_consumed = consumed;
-
-                sleep_unless_stopped(&stop_flag, Duration::from_millis(200));
-            }
-        });
-        self.position_timer_handle = Some(handle);
-    }
-
-    /// 停止位置推送定时器，等待线程退出
-    fn stop_position_timer(&mut self) {
-        if let Some(flag) = self.position_timer_stop.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(handle) = self.position_timer_handle.take() {
-            let _ = handle.join();
-        }
-    }
-
-    /// 启动 FFT 推送定时器（独立线程，每 50ms 推送一次频谱数据）
-    fn start_fft_timer(&mut self) {
-        self.stop_fft_timer();
-        if !self.fft_enabled() {
-            return;
-        }
-
-        let stop_flag = Arc::new(AtomicBool::new(false));
-        self.fft_timer_stop = Some(Arc::clone(&stop_flag));
-
-        let fft_enabled = Arc::clone(&self.fft_enabled);
-        let fft = Arc::clone(&self.fft);
-        let cb = match &self.event_callback {
-            Some(cb) => Arc::clone(cb),
-            None => return,
-        };
-
-        let handle = thread::spawn(move || {
-            while !stop_flag.load(Ordering::Relaxed) {
-                if fft_enabled.load(Ordering::Relaxed) {
-                    let (ldata, rdata) = fft.analyze();
-                    cb(PlayerEvent::FftData { ldata, rdata });
-                }
-                sleep_unless_stopped(&stop_flag, Duration::from_millis(50));
-            }
-        });
-        self.fft_timer_handle = Some(handle);
-    }
-
-    /// 停止 FFT 推送定时器，等待线程退出
-    fn stop_fft_timer(&mut self) {
-        if let Some(flag) = self.fft_timer_stop.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(handle) = self.fft_timer_handle.take() {
-            let _ = handle.join();
-        }
-    }
-
-    /// 设置 FFT 推送开关
-    pub fn set_fft_enabled(&mut self, enabled: bool) {
-        self.fft_enabled.store(enabled, Ordering::Relaxed);
-        self.fft.set_enabled(enabled);
-        if enabled && self.state == PlayerState::Playing {
-            self.start_fft_timer();
-        } else if !enabled {
-            self.stop_fft_timer();
-        }
-    }
-
-    /// 获取 FFT 推送开关状态
-    pub fn fft_enabled(&self) -> bool {
-        self.fft_enabled.load(Ordering::Relaxed)
-    }
-
-    /// 仅重建输出设备（系统休眠唤醒、设备热拔插等场景调用）
-    ///
-    /// 通过赋值新的 `AudioOutput` 触发旧 `AudioOutput` 的 `Drop`：旧 owner 线程
-    /// 退出 + 旧 `cpal::Stream` 在该线程内释放，再创建新 `AudioOutput`（新 owner 线程）。
-    /// 此函数仅替换输出设备，真正的状态恢复交由 `lib.rs` 层的 `reinit_output` 异步完成。
-    pub fn recreate_output_device(&mut self) -> Result<()> {
-        info!(device = ?self.selected_device_name, "开始重建音频输出");
-        self.output.take();
-        self.output = Some(AudioOutput::new(self.selected_device_name.as_deref())?);
-        Ok(())
-    }
-
-    /// 设置封面缓存目录
-    pub fn set_cover_cache_dir(&mut self, dir: String) {
-        self.cover_cache_dir = Some(dir);
-    }
-
-    /// 暴露给 lib.rs：cover 缓存目录
-    pub fn cover_cache_dir(&self) -> Option<&str> {
-        self.cover_cache_dir.as_deref()
-    }
-
-    /// 暴露给 lib.rs：归一化开关
-    pub fn is_normalization_enabled(&self) -> bool {
-        self.normalization_enabled
-    }
-
-    /// 给 lib.rs async load 用：原子地发出停止信号 + take 所有旧线程 handle
-    /// 调用方负责在工作线程 join 这些 handle，主线程持锁阶段不阻塞
-    /// 返回旧输出流供工作线程在打开新流前释放；token 用于校验本次 load 是否已被取代
-    pub fn take_for_async_load(
-        &mut self,
-        handle: HttpCancelHandle,
-    ) -> (OldThreads, Option<AudioOutput>, u64) {
-        // 自增 token：本次 load 的标识；任何并发的更早 commit_loaded 比较时会发现不匹配
-        let token = self.load_token.fetch_add(1, Ordering::AcqRel) + 1;
-        if let Some(previous) = self.pending_load_handle.replace(handle) {
-            previous.cancel();
-        }
-
-        // 发停止信号（原子写，纳秒级）
-        if let Some(flag) = self.fade_cancel.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(flag) = self.position_timer_stop.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(flag) = self.fft_timer_stop.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-
-        if let Some(ref shared) = self.shared {
-            shared.stop();
-        }
-        if let Some(sink) = self.sink.take() {
-            sink.stop();
-        }
-        if let Some(ref shared) = self.shared {
-            shared.drain_buffer();
-        }
-        self.shared = None;
-        self.cover_raw = None;
-        self.seek_base = 0.0;
-        self.fft.reset();
-        self.equalizer.lock().reset_state();
-        self.tempo.lock().reset();
-
-        let old_threads = OldThreads {
-            decoder_thread: self.decoder_thread.take(),
-            position_timer: self.position_timer_handle.take(),
-            fft_timer: self.fft_timer_handle.take(),
-            fade_handle: self.fade_handle.take(),
-        };
-        (old_threads, self.output.take(), token)
-    }
-
-    /// token 是否仍是最新值（seek 失败回退到 load 前校验，避免复活已被取代的旧源）
-    pub fn is_load_token_current(&self, token: u64) -> bool {
-        token == self.load_token.load(Ordering::Acquire)
-    }
-
-    /// 获取加载代次，用于工作线程在创建输出流前快速放弃已过期任务
-    pub fn load_token_handle(&self) -> Arc<AtomicU64> {
-        Arc::clone(&self.load_token)
-    }
-
-    /// 获取解码侧共享均衡器
-    pub fn equalizer_handle(&self) -> Arc<Mutex<Equalizer>> {
-        Arc::clone(&self.equalizer)
-    }
-
-    /// 获取解码侧共享变速处理器
-    pub fn tempo_handle(&self) -> Arc<Mutex<StretchProcessor>> {
-        Arc::clone(&self.tempo)
-    }
-
-    /// 清理仍属于指定 load 的网络中断句柄
-    pub fn clear_pending_load(&mut self, token: u64) {
-        if self.is_load_token_current(token) {
-            self.pending_load_handle = None;
-        }
-    }
-
-    /// 给 lib.rs async seek 用：原子发出停止信号 + take 所有旧线程 handle（不 join）
-    ///
-    /// 返回 None 表示当前没有解码线程（空闲 / 已停止 / 正在异步加载被 load 取走），
-    /// 此时不做任何副作用——尤其不能 bump token，否则会误杀在途的 load
-    pub fn take_for_async_seek(&mut self) -> Option<SeekTake> {
-        self.decoder_thread.as_ref()?;
-
-        // 与 load 共用同一 token 序列：commit_seeked 时比对，防止 seek 期间发生的
-        // load/stop 完成后被本次 seek 的 commit 覆盖（旧曲复活 + 新解码线程泄漏）
-        let token = self.load_token.fetch_add(1, Ordering::AcqRel) + 1;
-
-        if let Some(flag) = self.fade_cancel.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(flag) = self.position_timer_stop.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-        if let Some(flag) = self.fft_timer_stop.take() {
-            flag.store(true, Ordering::Relaxed);
-        }
-
-        if let Some(ref shared) = self.shared {
-            shared.stop();
-        }
-        if let Some(sink) = self.sink.take() {
-            sink.stop();
-        }
-
-        let old_threads = OldThreads {
-            decoder_thread: self.decoder_thread.take(),
-            position_timer: self.position_timer_handle.take(),
-            fft_timer: self.fft_timer_handle.take(),
-            fade_handle: self.fade_handle.take(),
-        };
-
-        let (norm_enabled, norm_gain) = match self.shared.take() {
-            Some(s) => {
-                s.drain_buffer();
-                (s.is_normalization_enabled(), s.normalization_gain())
-            }
-            None => (self.normalization_enabled, 0.0),
-        };
-
-        self.fft.reset();
-
-        Some(SeekTake {
-            old_threads,
-            normalization_enabled: norm_enabled,
-            normalization_gain: norm_gain,
-            current_source: self.current_source.clone(),
-            was_playing: self.state == PlayerState::Playing,
-            output_sample_rate: self.output_sample_rate(),
-            token,
-            equalizer: Arc::clone(&self.equalizer),
-            tempo: Arc::clone(&self.tempo),
-        })
-    }
-
-    /// seek 三段式的最后一段：主线程持锁，attach 新 sink + 新解码线程
-    ///
-    /// 返回 false 表示本次 seek 已被更新的 load/seek/stop 取代，结果被丢弃
-    pub fn commit_seeked(
-        &mut self,
-        token: u64,
-        position_secs: f64,
-        shared: Arc<Shared>,
-        handle: JoinHandle<decoder::DecoderData>,
-    ) -> Result<bool> {
-        // 抢占检查：与 commit_loaded 同款，不一致则丢弃本次 seek 结果
-        if token != self.load_token.load(Ordering::Acquire) {
-            shared.stop();
-            // 解码线程读到 stop 信号后自行退出，故意不 join 避免阻塞主线程持锁阶段
-            drop(handle);
-            return Ok(false);
-        }
-
-        let sink = {
-            let output = self.ensure_output()?;
-            Arc::new(RodioPlayer::connect_new(output.mixer()))
-        };
-
-        let sample_rate = shared.sample_rate();
-        let decoder_source = DecoderSource::new(
-            Arc::clone(&shared),
-            Arc::clone(&self.fft),
-            sample_rate,
-            self.audio_channels,
-        );
-
-        let was_paused = self.state == PlayerState::Paused;
-        sink.set_volume(self.target_volume);
-        if was_paused {
-            sink.pause();
-        }
-        sink.append(decoder_source);
-
-        self.sink = Some(sink);
-        self.shared = Some(shared);
-        self.decoder_thread = Some(handle);
-        self.audio_sample_rate = sample_rate;
-        self.seek_base = position_secs;
-
-        if was_paused {
-            self.state = PlayerState::Paused;
-            self.emit(PlayerEvent::StateChanged {
-                state: PlayerState::Paused,
-            });
-        } else {
-            self.state = PlayerState::Playing;
-            self.emit(PlayerEvent::StateChanged {
-                state: PlayerState::Playing,
-            });
-            self.start_position_timer();
-            self.start_fft_timer();
-        }
-
-        Ok(true)
-    }
-
-    /// load 的下半部分：lib.rs async load 在 IO 完成后由主线程持锁调用
-    ///
-    /// `token` 为 take_for_async_load 时拿到的标识。本函数比对当前最新 token：
-    /// - 不一致 → 本次 load 已被更新的 load 抢占，丢弃 sink/shared，stop 解码线程后返回 None
-    /// - 一致 → 正常 attach 新资源
-    pub fn commit_loaded(
-        &mut self,
-        token: u64,
-        source: &str,
-        auto_play: bool,
-        loaded: LoadedPlayback,
-    ) -> Result<Option<AudioMetadata>> {
-        let LoadedPlayback {
-            mut metadata,
-            decode_handle,
-            shared,
-            output,
-            cancel,
-        } = loaded;
-        // 抢占检查：比对最新 token，不等说明已有更新的 load 在路上 / 已 commit
-        if token != self.load_token.load(Ordering::Acquire) {
-            if let Some(h) = cancel {
-                h.cancel();
-            }
-            // 停止新解码线程（它会写入 shared 但没人消费），让 join 能尽快返回
-            shared.stop();
-            // shared / sink / decode_handle 在此函数返回时 drop；解码线程读到 stop 信号后退出
-            // decode_handle 故意不 join，避免阻塞主线程持锁阶段（让解码线程在后台自然结束）
-            drop(decode_handle);
-            return Ok(None);
-        }
-
-        self.pending_load_handle = cancel;
-        self.output = Some(output);
-
-        let sink = {
-            let output = self.ensure_output()?;
-            Arc::new(RodioPlayer::connect_new(output.mixer()))
-        };
-
-        let decoder_source = DecoderSource::new(
-            Arc::clone(&shared),
-            Arc::clone(&self.fft),
-            metadata.sample_rate,
-            metadata.channels,
-        );
-
-        sink.set_volume(self.target_volume);
-        if !auto_play {
-            sink.pause();
-        }
-        sink.append(decoder_source);
-
-        self.sink = Some(sink);
-        self.shared = Some(shared);
-        self.decoder_thread = Some(decode_handle);
-        self.seek_base = 0.0;
-        self.current_source = Some(source.to_string());
-
-        self.audio_sample_rate = metadata.sample_rate;
-        self.audio_channels = metadata.channels;
-        self.audio_duration = metadata.duration_secs;
-        self.cover_raw = metadata.cover_raw.take();
-
-        if auto_play {
-            self.state = PlayerState::Playing;
-            self.emit(PlayerEvent::StateChanged {
-                state: PlayerState::Playing,
-            });
-            self.start_position_timer();
-            self.start_fft_timer();
-        } else {
-            self.state = PlayerState::Paused;
-            self.emit(PlayerEvent::StateChanged {
-                state: PlayerState::Paused,
-            });
-        }
-
-        Ok(Some(metadata))
-    }
-
-    /// 恢复播放。Paused 时渐入恢复；Stopped/Idle/已播完时返回 Some(source)，
-    /// 由 lib.rs 走 async load 复活——网络源的打开可达数秒，不能在锁内同步执行
-    pub fn play(&mut self) -> Result<Option<String>> {
-        // 如果当前在"播放"状态但实际已结束，先标记为停止
-        // 此时解码线程已自然退出，stop_internal 的 join 立即返回，不会阻塞
-        if self.state == PlayerState::Playing && self.is_finished() {
-            self.stop_internal();
-            self.state = PlayerState::Stopped;
-        }
-
-        match self.state {
-            // 已经在播放且未结束，忽略
-            PlayerState::Playing => Ok(None),
-            // 暂停状态：渐入恢复
-            PlayerState::Paused => {
-                // 先取消未完成的渐出：否则其完成回调可能在 sink.play() 之后执行
-                // sink.pause()，导致状态 Playing 但实际无声
-                self.cancel_fade();
-                if let Some(ref sink) = self.sink {
-                    sink.set_volume(0.0);
-                    sink.play();
-                }
-
-                self.state = PlayerState::Playing;
-                self.emit(PlayerEvent::StateChanged {
-                    state: PlayerState::Playing,
-                });
-                self.start_position_timer();
-                self.start_fft_timer();
-
-                // 非阻塞渐入
-                self.start_fade(0.0, self.target_volume, None);
-                Ok(None)
-            }
-            // 停止/空闲/播放结束：交给调用方异步从头重新加载
-            PlayerState::Stopped | PlayerState::Idle => Ok(self.current_source.clone()),
-        }
-    }
-
-    /// 暂停播放（非阻塞渐出，渐出完成后 sink.pause）
-    pub fn pause(&mut self) {
-        if self.state != PlayerState::Playing {
-            return;
-        }
-
-        // 先切换状态并发射事件，让前端立即响应
-        self.state = PlayerState::Paused;
-        self.emit(PlayerEvent::StateChanged {
-            state: PlayerState::Paused,
-        });
-
-        // 立即启动非阻塞渐出，避免被后续 stop_*_timer 的 join 阻塞
-        // fade 完成后在回调中执行 sink.pause + 恢复音量
-        let target_volume = self.target_volume;
-        let sink_for_callback = self.sink.as_ref().map(Arc::clone);
-        self.start_fade(
-            target_volume,
-            0.0,
-            Some(Box::new(move || {
-                if let Some(sink) = sink_for_callback {
-                    sink.pause();
-                    sink.set_volume(target_volume);
-                }
-            })),
-        );
-
-        // 渐出已在后台运行，再同步停止定时器（join 开销不会影响音频淡出时序）
-        self.stop_position_timer();
-        self.stop_fft_timer();
-    }
-
-    /// 停止播放并释放资源
-    /// 显式停止：清掉 current_source，避免后续 play() 在 Stopped 态下用残留源复活上一首
-    /// （`stop_internal` 是内部过渡用，不清；load() 会立即用新源覆盖）
-    pub fn stop(&mut self) {
-        // 使在途的 async load/seek 在 commit 时被拒绝，防止 stop 后被复活
-        self.load_token.fetch_add(1, Ordering::AcqRel);
-        if let Some(handle) = self.pending_load_handle.take() {
-            handle.cancel();
-        }
-        self.stop_internal();
-        self.current_source = None;
-        self.state = PlayerState::Stopped;
-        self.emit(PlayerEvent::StateChanged {
-            state: PlayerState::Stopped,
-        });
-    }
-
-    fn stop_internal(&mut self) {
-        // 1. 取消渐变并等待渐变线程退出（释放 Arc<Sink>）
-        self.cancel_fade();
-        // 2. 停止定时器并等待线程退出（释放 Arc<Shared> 和 Arc<EventEmitter>）
-        self.stop_position_timer();
-        self.stop_fft_timer();
-        // 3. 通知解码线程停止
-        if let Some(ref shared) = self.shared {
-            shared.stop();
-        }
-        // 4. 释放 Sink（drop DecoderSource，解除迭代器阻塞）
-        if let Some(sink) = self.sink.take() {
-            sink.stop();
-        }
-        // 5. 等待解码线程退出，回收 DecoderData（FFmpeg 资源在此 drop）
-        if let Some(handle) = self.decoder_thread.take() {
-            let _ = handle.join();
-        }
-        // 6. 清空共享缓冲区（即使还有外部 Arc 引用，缓冲区数据也立即释放）
-        if let Some(ref shared) = self.shared {
-            shared.drain_buffer();
-        }
-        self.shared = None;
-        self.cover_raw = None;
-        self.seek_base = 0.0;
-    }
-
-    /// 设置音量（0.0 ~ 1.0）
-    pub fn set_volume(&mut self, volume: f32) {
-        self.target_volume = volume;
-        if let Some(ref sink) = self.sink {
-            sink.set_volume(volume);
-        }
-    }
-
-    /// 获取当前音量
-    pub fn volume(&self) -> f32 {
-        self.target_volume
-    }
-
-    /// 设置渐变时长（毫秒），0 表示禁用渐变
-    pub fn set_fade_duration(&mut self, duration_ms: u64) {
-        self.fade_duration_ms = duration_ms;
-    }
-
-    /// 获取渐变时长（毫秒）
-    pub fn fade_duration(&self) -> u64 {
-        self.fade_duration_ms
-    }
-
-    /// 获取当前播放位置（秒），基于实际消费的采样数
-    pub fn position(&self) -> f64 {
-        match &self.shared {
-            Some(shared) => self.seek_base + shared.consumed_position(),
-            None => self.seek_base,
-        }
-    }
-
-    /// 获取总时长（秒）
-    pub fn duration(&self) -> f64 {
-        self.audio_duration
-    }
-
-    /// 获取当前播放状态
-    pub fn state(&self) -> PlayerState {
-        self.state
-    }
-
-    /// 获取 FFT 频谱数据（128 个频段）
-    pub fn fft_data(&self) -> (Vec<f32>, Vec<f32>) {
-        self.fft.analyze()
-    }
-
-    /// 获取缓存的原始封面数据（load 时一次性提取）
-    pub fn cover_raw(&self) -> Option<&[u8]> {
-        self.cover_raw.as_deref()
-    }
-
-    /// 检查播放是否已结束
-    pub fn is_finished(&self) -> bool {
-        match (&self.shared, &self.sink) {
-            (Some(shared), Some(sink)) => shared.is_done() && sink.empty(),
-            _ => false,
-        }
-    }
-
-    /// 设置音量归一化开关
-    pub fn set_normalization_enabled(&mut self, enabled: bool) {
-        self.normalization_enabled = enabled;
-        if let Some(ref shared) = self.shared {
-            shared.set_normalization_enabled(enabled);
-        }
-    }
-
-    /// 获取音量归一化开关状态
-    pub fn normalization_enabled(&self) -> bool {
-        self.normalization_enabled
-    }
-
-    /// 设置均衡器开关
-    pub fn set_equalizer_enabled(&mut self, enabled: bool) {
-        self.equalizer.lock().set_enabled(enabled);
-    }
-
-    /// 获取均衡器开关状态
-    pub fn equalizer_enabled(&self) -> bool {
-        self.equalizer.lock().enabled()
-    }
-
-    /// 更新所有频段增益（dB），长度需为 EQ_BAND_COUNT
-    pub fn set_equalizer_bands(&mut self, gains_db: &[f32]) {
-        self.equalizer.lock().set_band_gains(gains_db);
-    }
-
-    /// 获取所有频段当前增益（dB）
-    pub fn equalizer_bands(&self) -> [f32; EQ_BAND_COUNT] {
-        self.equalizer.lock().band_gains_db()
-    }
-
-    /// 设置前级增益（dB，自动 clamp 到 ±12）
-    pub fn set_preamp_gain(&mut self, db: f32) {
-        self.equalizer.lock().set_preamp_db(db);
-    }
-
-    /// 获取前级增益（dB）
-    pub fn preamp_gain(&self) -> f32 {
-        self.equalizer.lock().preamp_db()
-    }
-
-    /// 设置播放速度（自动 clamp 到 [0.5, 2.0]）
-    pub fn set_speed(&mut self, speed: f32) {
-        self.tempo.lock().set_speed(speed);
-    }
-
-    /// 设置音调偏移（半音，自动 clamp 到 [-12, 12]）。
-    /// sync=ON 时立即下发；sync=OFF 时只更新内部值，不影响声音
-    pub fn set_pitch(&mut self, semitones: i8) {
-        self.tempo.lock().set_pitch(semitones);
-    }
-
-    /// 设置"音调同步"开关（true = 变速保音调，默认）
-    pub fn set_pitch_sync(&mut self, sync: bool) {
-        self.tempo.lock().set_pitch_sync(sync);
-    }
-
-    /// 获取当前播放速度
-    pub fn speed(&self) -> f32 {
-        self.tempo.lock().speed()
-    }
-
-    /// 获取当前音调（半音）
-    pub fn pitch(&self) -> i8 {
-        self.tempo.lock().pitch()
-    }
-
-    /// 获取"音调同步"开关
-    pub fn pitch_sync(&self) -> bool {
-        self.tempo.lock().pitch_sync()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn decode_failure_mid_stream_emits_source_error() {
-        let shared = Shared::new(48_000, 2);
-        shared.mark_decode_failed();
-
-        assert!(matches!(
-            playback_completion_event(&shared, 120.0, 30.0),
-            PlayerEvent::SourceError
-        ));
-    }
-
-    #[test]
-    fn decode_failure_near_end_is_treated_as_ended() {
-        let shared = Shared::new(48_000, 2);
-        shared.mark_decode_failed();
-
-        assert!(matches!(
-            playback_completion_event(&shared, 120.0, 118.0),
-            PlayerEvent::Ended
-        ));
-    }
-
-    #[test]
-    fn unknown_duration_failure_emits_source_error() {
-        let shared = Shared::new(48_000, 2);
-        shared.mark_decode_failed();
-
-        assert!(matches!(
-            playback_completion_event(&shared, 0.0, 30.0),
-            PlayerEvent::SourceError
-        ));
-    }
-}
